requirements.txt
numpy==1.16.4
pandas==0.24.2
pdf2image==1.5.4
Pillow==6.0.0
PyPDF2==1.26.0
pytesseract==0.2.6
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
